content-header and icons and actions

+++ sh-server/build/database/database.js
@@ -1,9 +1,10 @@
 "use strict";
 var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, generator) {
+    function adopt(value) { return value instanceof P ? value : new P(function (resolve) { resolve(value); }); }
     return new (P || (P = Promise))(function (resolve, reject) {
         function fulfilled(value) { try { step(generator.next(value)); } catch (e) { reject(e); } }
         function rejected(value) { try { step(generator["throw"](value)); } catch (e) { reject(e); } }
-        function step(result) { result.done ? resolve(result.value) : new P(function (resolve) { resolve(result.value); }).then(fulfilled, rejected); }
+        function step(result) { result.done ? resolve(result.value) : adopt(result.value).then(fulfilled, rejected); }
         step((generator = generator.apply(thisArg, _arguments || [])).next());
     });
 };
@@ -312,12 +313,17 @@ class DbContext {
             else {
                 whereConditions.push("1 = 1");
             }
-            let deleteQuery = `DELETE FROM ${tableName} WHERE ${whereConditions.join(' OR ')}`;
-            this.connection.query(deleteQuery, (err, result) => {
-                if (err)
-                    reject(err);
-                resolve(result);
-            });
+            if (whereConditions.length == 0) {
+                resolve();
+            }
+            else {
+                let deleteQuery = `DELETE FROM ${tableName} WHERE ${whereConditions.join(' OR ')}`;
+                this.connection.query(deleteQuery, (err, result) => {
+                    if (err)
+                        reject(err);
+                    resolve(result);
+                });
+            }
         });
         return deletePromise;
     }
